+++ app/Http/Controllers/Admin/HospitalController.php
@@ -23,6 +23,7 @@ class HospitalController extends Controller
 
             return datatables()->of($data)
                 ->addColumn('action', function ($data) {
+                    $user            = auth()->user();
                     $editRoute       = 'admin.hospitals.edit';
                     $deleteRoute     = 'admin.hospitals.destroy';
                     $viewRoute       = 'admin.hospitals.show';
@@ -31,16 +32,20 @@ class HospitalController extends Controller
 
                     $action = "";
 
-                    $action .= '
-                        <a class="btn btn-warning" id="btn-edit" type="button" data-url="' . route($editRoute, $dataId) . '">
-                            <i class="fe fe-pencil"></i>
-                        </a> ';
+                    if ($user->can('update hospitals')) {
+                        $action .= '
+                            <a class="btn btn-warning" id="btn-edit" type="button" data-url="' . route($editRoute, $dataId) . '">
+                                <i class="fe fe-pencil"></i>
+                            </a> ';
+                    }
 
-                    $action .= '
-                        <button class="btn btn-danger delete-item" 
-                            data-label="' . $dataDeleteLabel . '" data-url="' . route($deleteRoute, $dataId) . '">
-                            <i class="fe fe-trash"></i>
-                        </button> ';
+                    if ($user->can('delete hospitals')) {
+                        $action .= '
+                            <button class="btn btn-danger delete-item" 
+                                data-label="' . $dataDeleteLabel . '" data-url="' . route($deleteRoute, $dataId) . '">
+                                <i class="fe fe-trash"></i>
+                            </button> ';
+                    }
 
                     $group = '<div class="btn-group btn-group-sm mb-1 mb-md-0" role="group">
                         ' . $action . '
@@ -65,7 +70,11 @@ class HospitalController extends Controller
      */
     public function create()
     {
-        //
+        $user = auth()->user();
+
+        if (!$user->can('create hospitals')) {
+            abort(403);
+        }
     }
 
     /**
@@ -73,54 +82,60 @@ class HospitalController extends Controller
      */
     public function store(Request $request)
     {
-        $validator = Validator::make([
-            'name'                  => $request->name,
-            'address'               => $request->address,
-            'phone'                 => $request->phone,
-            'whatsapp'              => $request->whatsapp,
-            'email'                 => $request->email,
-            'instagram'             => $request->instagram,
-            'facebook'              => $request->facebook,
-            'logo'                  => $request->logo
-        ], [
-            'name'                  => 'required|max:100|min:5|unique:hospitals,name,NULL,id',
-            'address'               => 'required|min:5',
-            'phone'                 => 'required|min:7',
-            'whatsapp'              => 'required|min:9',
-            'email'                 => 'required|email|unique:hospitals,email',
-            'logo'                  => 'required|mimes:png|max:1000',
-        ]);
+        $user = auth()->user();
 
-        if ($validator->fails()) {
-            return response()->json([
-                'status' => 400,
-                'errors' => $validator->messages(),
+        if ($user->can('create hospitals')) {
+            $validator = Validator::make([
+                'name'                  => $request->name,
+                'address'               => $request->address,
+                'phone'                 => $request->phone,
+                'whatsapp'              => $request->whatsapp,
+                'email'                 => $request->email,
+                'instagram'             => $request->instagram,
+                'facebook'              => $request->facebook,
+                'logo'                  => $request->logo
+            ], [
+                'name'                  => 'required|max:100|min:5|unique:hospitals,name,NULL,id',
+                'address'               => 'required|min:5',
+                'phone'                 => 'required|min:7',
+                'whatsapp'              => 'required|min:9',
+                'email'                 => 'required|email|unique:hospitals,email',
+                'logo'                  => 'required|mimes:png|max:1000',
             ]);
-        } else {
-            DB::beginTransaction();
-            try {
-                Hospital::create([
-                    'name'                  => $request->name,
-                    'address'               => $request->address,
-                    'phone'                 => $request->phone,
-                    'whatsapp'              => $request->whatsapp,
-                    'email'                 => $request->email,
-                    'instagram'             => $request->instagram,
-                    'facebook'              => $request->facebook,
-                    'logo'                  => request('logo') ? $request->file('logo')->store('images/hospital/logo') : null,
+
+            if ($validator->fails()) {
+                return response()->json([
+                    'status' => 400,
+                    'errors' => $validator->messages(),
                 ]);
-                DB::commit();
-                return response()->json([
-                    'status'  => 200,
-                    'message' => 'Hospital has been created',
-                ], 200);
-            } catch (Throwable $th) {
-                DB::rollBack();
-                return response()->json([
-                    'status'  => 500,
-                    'message' => $th->getMessage(),
-                ], 500);
+            } else {
+                DB::beginTransaction();
+                try {
+                    Hospital::create([
+                        'name'                  => $request->name,
+                        'address'               => $request->address,
+                        'phone'                 => $request->phone,
+                        'whatsapp'              => $request->whatsapp,
+                        'email'                 => $request->email,
+                        'instagram'             => $request->instagram,
+                        'facebook'              => $request->facebook,
+                        'logo'                  => request('logo') ? $request->file('logo')->store('images/hospital/logo') : null,
+                    ]);
+                    DB::commit();
+                    return response()->json([
+                        'status'  => 200,
+                        'message' => 'Hospital has been created',
+                    ], 200);
+                } catch (Throwable $th) {
+                    DB::rollBack();
+                    return response()->json([
+                        'status'  => 500,
+                        'message' => $th->getMessage(),
+                    ], 500);
+                }
             }
+        } else {
+            abort(403);
         }
     }
 
@@ -137,26 +152,32 @@ class HospitalController extends Controller
      */
     public function edit($id)
     {
-        try {
-            $id = Crypt::decryptString($id);
-            $data = Hospital::find($id);
+        $user = auth()->user();
 
-            if ($data) {
+        if ($user->can('update hospitals')) {
+            try {
+                $id = Crypt::decryptString($id);
+                $data = Hospital::find($id);
+
+                if ($data) {
+                    return response()->json([
+                        'status'    => 200,
+                        'data'      => $data,
+                    ]);
+                } else {
+                    return response()->json([
+                        'status'    => 404,
+                        'message'   => 'Hospital Not Found',
+                    ]);
+                }
+            } catch (\Throwable $e) {
                 return response()->json([
-                    'status'    => 200,
-                    'data'      => $data,
-                ]);
-            } else {
-                return response()->json([
-                    'status'    => 404,
-                    'message'   => 'Hospital Not Found',
-                ]);
+                    'status'  => 500,
+                    'message' => "Error on line {$e->getLine()}: {$e->getMessage()}",
+                ], 500);
             }
-        } catch (\Throwable $e) {
-            return response()->json([
-                'status'  => 500,
-                'message' => "Error on line {$e->getLine()}: {$e->getMessage()}",
-            ], 500);
+        } else {
+            abort(403);
         }
     }
 
@@ -165,79 +186,85 @@ class HospitalController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $data = Hospital::find($id);
+        $user = auth()->user();
 
-        $validator = Validator::make([
-            'name'                  => $request->name,
-            'address'               => $request->address,
-            'phone'                 => $request->phone,
-            'whatsapp'              => $request->whatsapp,
-            'email'                 => $request->email,
-            'instagram'             => $request->instagram,
-            'facebook'              => $request->facebook,
-            'logo'                  => $request->logo,
-        ], [
-            'name'                  => 'required|max:100|min:5|unique:hospitals,name,' . $data->id,
-            'address'               => 'required|min:5',
-            'phone'                 => 'required|min:7',
-            'whatsapp'              => 'required|min:9',
-            'email'                 => 'required|email|unique:hospitals,email,' . $data->id,
-            'logo'                  => request('logo') ? 'mimes:png|max:1000' : '',
-        ]);
+        if ($user->can('update hospitals')) {
+            $data = Hospital::find($id);
 
-        if ($validator->fails()) {
-            return response()->json([
-                'status' => 400,
-                'errors' => $validator->messages(),
+            $validator = Validator::make([
+                'name'                  => $request->name,
+                'address'               => $request->address,
+                'phone'                 => $request->phone,
+                'whatsapp'              => $request->whatsapp,
+                'email'                 => $request->email,
+                'instagram'             => $request->instagram,
+                'facebook'              => $request->facebook,
+                'logo'                  => $request->logo,
+            ], [
+                'name'                  => 'required|max:100|min:5|unique:hospitals,name,' . $data->id,
+                'address'               => 'required|min:5',
+                'phone'                 => 'required|min:7',
+                'whatsapp'              => 'required|min:9',
+                'email'                 => 'required|email|unique:hospitals,email,' . $data->id,
+                'logo'                  => request('logo') ? 'mimes:png|max:1000' : '',
             ]);
-        } else {
-            if ($data) {
-                DB::beginTransaction();
 
-                //--Membuat kondisi langsung mendelete gambar yang lama pada storage
-                if (request('logo')) {
-                    if ($data->logo) {
-                        Storage::delete($data->logo);
-                    }
-                    $logo = request()->file('logo')->store('images/hospital/logo');
-                } elseif ($data->logo) {
-                    $logo = $data->logo;
-                } else {
-                    $logo = null;
-                }
-                //--End
-
-                try {
-                    $data->update([
-                        'name'                  => $request->name,
-                        'address'               => $request->address,
-                        'phone'                 => $request->phone,
-                        'whatsapp'              => $request->whatsapp,
-                        'email'                 => $request->email,
-                        'instagram'             => $request->instagram,
-                        'facebook'              => $request->facebook,
-                        'logo'                  => $logo
-                    ]);
-
-                    DB::commit();
-
-                    return response()->json([
-                        'status'  => 200,
-                        'message' => 'Hospital has been updated',
-                    ], 200);
-                } catch (\Throwable $th) {
-                    DB::rollBack();
-                    return response()->json([
-                        'status'  => 500,
-                        'message' => $th->getMessage(),
-                    ], 500);
-                }
-            } else {
+            if ($validator->fails()) {
                 return response()->json([
-                    'status' => 404,
-                    'message' => 'Data not found..!',
+                    'status' => 400,
+                    'errors' => $validator->messages(),
                 ]);
+            } else {
+                if ($data) {
+                    DB::beginTransaction();
+
+                    //--Membuat kondisi langsung mendelete gambar yang lama pada storage
+                    if (request('logo')) {
+                        if ($data->logo) {
+                            Storage::delete($data->logo);
+                        }
+                        $logo = request()->file('logo')->store('images/hospital/logo');
+                    } elseif ($data->logo) {
+                        $logo = $data->logo;
+                    } else {
+                        $logo = null;
+                    }
+                    //--End
+
+                    try {
+                        $data->update([
+                            'name'                  => $request->name,
+                            'address'               => $request->address,
+                            'phone'                 => $request->phone,
+                            'whatsapp'              => $request->whatsapp,
+                            'email'                 => $request->email,
+                            'instagram'             => $request->instagram,
+                            'facebook'              => $request->facebook,
+                            'logo'                  => $logo
+                        ]);
+
+                        DB::commit();
+
+                        return response()->json([
+                            'status'  => 200,
+                            'message' => 'Hospital has been updated',
+                        ], 200);
+                    } catch (\Throwable $th) {
+                        DB::rollBack();
+                        return response()->json([
+                            'status'  => 500,
+                            'message' => $th->getMessage(),
+                        ], 500);
+                    }
+                } else {
+                    return response()->json([
+                        'status' => 404,
+                        'message' => 'Data not found..!',
+                    ]);
+                }
             }
+        } else {
+            abort(403);
         }
     }
 
@@ -246,38 +273,44 @@ class HospitalController extends Controller
      */
     public function destroy($id)
     {
-        DB::beginTransaction();
-        try {
-            $id = Crypt::decryptString($id);
-            $data = Hospital::find($id);
+        $user = auth()->user();
 
-            if (!$data) {
+        if ($user->can('delete hospitals')) {
+            DB::beginTransaction();
+            try {
+                $id = Crypt::decryptString($id);
+                $data = Hospital::find($id);
+
+                if (!$data) {
+                    return response()->json([
+                        'status'  => 404,
+                        'message' => "Data not found!",
+                    ], 404);
+                }
+
+                //Kondisi apabila terdapat path gambar pada tabel
+                if ($data->logo != null) {
+                    Storage::delete($data->logo);
+                    $data->delete();
+                } else {
+                    $data->delete();
+                }
+                //End kondisi
+
+                DB::commit();
+
                 return response()->json([
-                    'status'  => 404,
-                    'message' => "Data not found!",
-                ], 404);
+                    'status'  => 200,
+                    'message' => "Hospital has been deleted..!",
+                ], 200);
+            } catch (\Throwable $e) {
+                return response()->json([
+                    'status'  => 500,
+                    'message' => "Error on line {$e->getLine()}: {$e->getMessage()}",
+                ], 500);
             }
-
-            //Kondisi apabila terdapat path gambar pada tabel
-            if ($data->logo != null) {
-                Storage::delete($data->logo);
-                $data->delete();
-            } else {
-                $data->delete();
-            }
-            //End kondisi
-
-            DB::commit();
-
-            return response()->json([
-                'status'  => 200,
-                'message' => "Hospital has been deleted..!",
-            ], 200);
-        } catch (\Throwable $e) {
-            return response()->json([
-                'status'  => 500,
-                'message' => "Error on line {$e->getLine()}: {$e->getMessage()}",
-            ], 500);
+        } else {
+            abort(403);
         }
     }
 }
